// src/app/api/models/icd-10.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles del ICD-10 Diagnosis a crear.
 */
export interface Icd10 {

  /**
   * Status of the diagnosis
   */
  active?: boolean;

  /**
   * ICD-10 Code
   */
  code: string;

  /**
   * Diagnosis Description
   */
  description: string;

  /**
   * Identificador único
   */
  id?: number;
}